move run_pad to test_pad
